Fixed font sizes

+++ spacearena-core/src/main/java/ru/spacearena/engine/util/FloatMathUtils.java
@@ -211,7 +211,7 @@ public class FloatMathUtils {
         return size(min, max) / 2;
     }
 
-    public static long round(float v) {
+    public static int round(float v) {
         return Math.round(v);
     }
